requirements.txt
backcall==0.1.0
Click==7.0
decorator==4.4.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
imgui==1.1.0
ipython==7.12.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.5.2
pyinotify==0.9.6
PyOpenGL==3.1.5
six==1.14.0
therepl==0.2.0
traitlets==4.3.3
wcwidth==0.1.8
Werkzeug==1.0.0
